some changes to the ui
