+++ app/styles/dashboard/data-explorer.scss
@@ -8,15 +8,13 @@
   max-width: 100%; 
   overflow: hidden;
 
-  &__value-type {
-    pointer-events: auto; 
-    margin-bottom: 15px;
-  }
 
   &__parameter-selection {
     display: flex;
     gap: 15px;
     pointer-events: auto;
+    flex-direction: column;
+    padding-right: 27px;
   }
 
   &__help-button {
